refactor: consolidate rawSvgPlugin and svg-test-resolver into libs/vitest-config
Both plugins were duplicated across apps:
- rawSvgPlugin: 3 copies (shelter-web, wildfires, betterangels-admin)
- vitest-svg-resolver: 2 copies (shelter-web, wildfires)

Moved both into libs/vitest-config/src/ and updated all three apps
to import from @monorepo/vitest-config. The svgTestResolverIfVitest
helper encapsulates the process.env.VITEST guard so app configs just
spread ...svgTestResolverIfVitest(__dirname).
