+++ libs/vitest-config/src/index.ts
@@ -9,3 +9,5 @@ export const baseVitestConfig = defineConfig({
   },
 });
 export { monorepoTsconfigAliases } from './monorepo-aliases';
+export { rawSvgPlugin } from './raw-svg-plugin';
+export { svgTestResolverIfVitest } from './svg-test-resolver';
